Fixed general bugs(Creating Directory on Windows, Converting format of vertexbuffer, Mouse with many buttons). Fixed x64 bugs(Comparing int of different type). Fixed x64 warnings. Improved devConfig.cmake for x64 builds.

+++ inc/core/lxIterator.h
@@ -449,7 +449,7 @@ public:
 	}
 	int operator-(StrideBaseIterator& other) const
 	{
-		return (m_Ptr - other.m_Ptr) / m_Stride;
+		return (int)(m_Ptr - other.m_Ptr) / m_Stride;
 	}
 
 	StrideBaseIterator& operator++()
